[FLINK-39392][table] Move BuiltInCondition out of TraitCondition for stricter encapsulation
